requirements.txt
# Dependencias del Proyecto Iris Classification
# Instalar con: pip install -r requirements.txt

# Core ML libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# Model persistence
joblib>=1.3.0

# Testing (opcional, para desarrollo)
pytest>=7.4.0
pytest-cov>=4.1.0
